requirements.txt
pyrogram
tgcrypto
yt-dlp @ git+https://github.com/yt-dlp/yt-dlp.git@master
deep-translator
aiohttp
requests
gallery-dl
img2pdf
Pillow